finish haaapy's documents and move some methods to right package
