+++ docs/source/discovery_methods/constraint/index.rst
@@ -1,2 +1,11 @@
 Constraint-based
 =================
+.. toctree::
+   :maxdepth: 2
+
+   PBSCM/pbscm
+
+.. toctree::
+   :maxdepth: 2
+
+   PBSCM_PGF/pbscm_pgf
