+++ docs/source/discovery_methods/functional_based/index.rst
@@ -1,2 +1,12 @@
 Functional-based
 =================
+
+.. toctree::
+   :maxdepth: 2
+
+   ANM/anm
+
+.. toctree::
+   :maxdepth: 2
+
+   SHP/shp
