+++ docs/source/effect_methods/index.rst
@@ -1,0 +1,5 @@
+Discovery methods
+=================
+
+In this section, we introduce effect methods implemented in CDMIR.
+
